feat: admin-create-magazine api개발 #82

+++ src/main/java/com/gam/api/dto/admin/magazine/request/MagazineCreateRequestDTO.java
@@ -1,6 +1,5 @@
 package com.gam.api.dto.admin.magazine.request;
 
-
 import lombok.Builder;
 
 import java.util.List;
@@ -13,16 +12,7 @@ public record MagazineCreateRequestDTO(
         List<QuestionVO> questions
 ) {
     @Builder
-    public
-    record QuestionVO (
-            Long questionOrder,
-            String question,
-            String answer,
-            String answerImage,
-            String imageCaption
-    ) {
-    }
-
+    public record QuestionVO( int questionOrder, String question, String answer, String answerImage, String imageCaption) {}
 }
